+++ src/main/java/library/presentation/GUI.java
@@ -2,7 +2,6 @@ package library.presentation;
 
 import library.model.implementation.*;
 import library.persistence.implementation.DatabaseFactory;
-import library.persistence.implementation.DatabaseService;
 
 import javax.swing.*;
 import javax.swing.text.BadLocationException;
@@ -70,6 +69,7 @@ public class GUI extends JFrame {
 
     private final ResourceBundle resourceBundle;
     private static final String RESOURCE_BUNDLE = "i18n/gui/gui"; //NON-NLS
+    DatabaseFactory factory = new DatabaseFactory(); // gui does not use db types from persistence but uses methods from the factory
 
 
     public GUI() {
@@ -145,7 +145,7 @@ public class GUI extends JFrame {
         spClosestRelative = new JScrollPane(tpClosestRelative);
         spVisits = new JScrollPane(tpVisits);
         spOther = new JScrollPane(tpOther);
-        spTextResident = new JScrollPane[DatabaseFactory.residents.size()];
+        spTextResident = new JScrollPane[factory.residents.size()];
 
         jpSpecific.add(spBaseData);
         jpSpecific.add(spMedication);
@@ -197,11 +197,11 @@ public class GUI extends JFrame {
 
     private void databaseConnectionForFilters() {
         shifts = new String[]{resourceBundle.getString("morning_shift"), resourceBundle.getString("late_shift"), resourceBundle.getString("night_shift")};
-        time = new String[(DatabaseFactory.shiftSchedules.size() / 3)];
+        time = new String[(factory.shiftSchedules.size() / 3)];
         int n = 0;
         String previous = null;
-        for (int i = 0; i < DatabaseFactory.shiftSchedules.size(); i++) {
-            Date date = DatabaseFactory.shiftSchedules.get(i).getDate();
+        for (int i = 0; i < factory.shiftSchedules.size(); i++) {
+            Date date = factory.shiftSchedules.get(i).getDate();
             SimpleDateFormat formatter = new SimpleDateFormat("dd.MM.yyyy");
             String dateInString = formatter.format(date);
             if (i != 0) {
@@ -297,9 +297,9 @@ public class GUI extends JFrame {
     }
 
     private void editButtonInitialization() {
-        btnEditResident = new JButton[DatabaseFactory.residents.size()];
+        btnEditResident = new JButton[factory.residents.size()];
 
-        for (int i = 0; i < DatabaseFactory.residents.size(); i++) {
+        for (int i = 0; i < factory.residents.size(); i++) {
             btnEditResident[i] = new JButton();
             btnEditResident[i].setBackground(lightgrey);
             btnEditResident[i].setBorder(BorderFactory.createLineBorder(Color.BLACK));
@@ -311,11 +311,11 @@ public class GUI extends JFrame {
     }
 
     private void residentTextAreaInitialization() {
-        taResident = new JTextArea[DatabaseFactory.residents.size()];
+        taResident = new JTextArea[factory.residents.size()];
 
-        for (int i = 0; i < DatabaseFactory.residents.size(); i++) {
-            int resID = DatabaseFactory.residents.get(i).getResID();
-            taResident[i] = new JTextArea(MessageFormat.format(resourceBundle.getString("incidents.0"), Objects.requireNonNull(Incident.get(resID)).getDescription()));
+        for (int i = 0; i < factory.residents.size(); i++) {
+            int resID = factory.residents.get(i).getResID();
+            taResident[i] = new JTextArea(MessageFormat.format(resourceBundle.getString("incidents.0"),factory.getSingleIncident(resID).getDescription())); //todo reicht es hier den ersten Incident zu nehmen oder sollte man eine Datumskontrolle machen ?
             taResident[i].setLineWrap(true);
             taResident[i].setWrapStyleWord(true);
             taResident[i].setBorder(BorderFactory.createLineBorder(Color.BLACK));
@@ -327,10 +327,10 @@ public class GUI extends JFrame {
     }
 
     private void roomLabelInitalization() {
-        lblRoom = new JLabel[DatabaseFactory.residents.size()];
+        lblRoom = new JLabel[factory.residents.size()];
 
-        for (int i = 0; i < DatabaseFactory.residents.size(); i++) {
-            lblRoom[i] = new JLabel(MessageFormat.format(resourceBundle.getString("room.0"), DatabaseFactory.residents.get(i).getRoom()), SwingConstants.CENTER);
+        for (int i = 0; i < factory.residents.size(); i++) {
+            lblRoom[i] = new JLabel(MessageFormat.format(resourceBundle.getString("room.0"), factory.residents.get(i).getRoom()), SwingConstants.CENTER);
             lblRoom[i].setBackground(lightgrey);
             lblRoom[i].setBorder(BorderFactory.createLineBorder(Color.BLACK));
             lblRoom[i].setOpaque(true);
@@ -341,10 +341,10 @@ public class GUI extends JFrame {
     }
 
     private void residentButtonInitialization() {
-        btnResident = new JButton[DatabaseFactory.residents.size()];
+        btnResident = new JButton[factory.residents.size()];
 
-        for (int i = 0; i < DatabaseFactory.residents.size(); i++) {
-            btnResident[i] = new JButton(DatabaseFactory.residents.get(i).getName() + " " + DatabaseFactory.residents.get(i).getSurname());
+        for (int i = 0; i < factory.residents.size(); i++) {
+            btnResident[i] = new JButton(factory.residents.get(i).getName() + " " + factory.residents.get(i).getSurname());
             btnResident[i].setBackground(lightgrey);
             btnResident[i].setBorder(BorderFactory.createLineBorder(Color.BLACK));
             btnResident[i].setPreferredSize(new Dimension(302, 51));
@@ -355,8 +355,8 @@ public class GUI extends JFrame {
     }
 
     public void setResidentIncidentText() {
-        for (int i = 0; i < DatabaseFactory.residents.size(); i++) {
-            int resID = DatabaseFactory.residents.get(i).getResID();
+        for (int i = 0; i < factory.residents.size(); i++) {
+            int resID = factory.residents.get(i).getResID();
             String dateString = (String) jcbTime.getSelectedItem(); //String format
             Date date = null;
             try {
@@ -364,7 +364,7 @@ public class GUI extends JFrame {
             } catch (ParseException e) {
                 e.printStackTrace();
             }
-            taResident[i].setText(MessageFormat.format(resourceBundle.getString("incidents.0"), Objects.requireNonNull(Incident.get(resID, date)).getDescription()));
+            taResident[i].setText(MessageFormat.format(resourceBundle.getString("incidents.0"),  factory.getSingleIncident(resID, date).getDescription()));
         }
     }
 
@@ -379,14 +379,14 @@ public class GUI extends JFrame {
         } catch (ParseException e) {
             e.printStackTrace();
         }
-        String shiftIncident = Objects.requireNonNull(ShiftSchedule.get(shiftCategory, date)).getShiftIncidents();
+        String shiftIncident = factory.getSingleShiftSchedule(shiftCategory, date).getShiftIncidents();
         taAll.setText(shiftIncident);
     }
 
     public void setResidentSpecificData(int index) {
-        Resident selectedResident = Resident.get(index);
-        MedPlan medPlan = MedPlan.get(selectedResident.getResID());
-        ICE ice = ICE.get(selectedResident.getResID());
+        Resident selectedResident = factory.getSingleResident(index);
+        MedPlan medPlan =factory.getSingleMedPlan(selectedResident.getResID());
+        ICE ice = factory.getSingleICE(selectedResident.getResID());
 
         setBaseData(selectedResident);
         setMedication(selectedResident, medPlan);
@@ -433,7 +433,7 @@ public class GUI extends JFrame {
             docMedication.insertString(docMedication.getLength(), resourceBundle.getString("medication.id"), attrSubHeader);
             docMedication.insertString(docMedication.getLength(), String.valueOf(medPlan.getMedicID()), attrText);
             docMedication.insertString(docMedication.getLength(), resourceBundle.getString("medication.name"), attrSubHeader);
-            docMedication.insertString(docMedication.getLength(), Medication.get(medPlan.getMedID()), attrText);
+            docMedication.insertString(docMedication.getLength(),  factory.getSingleMedication(medPlan.getMedID()), attrText);
 
         } catch (NullPointerException e) {
             System.out.println("NullPointerException");
@@ -481,7 +481,7 @@ public class GUI extends JFrame {
         try {
             tpVisits.setText(" ");
             docVisits.insertString(docVisits.getLength(), resourceBundle.getString("visits"), attrHeader);
-            docVisits.insertString(docVisits.getLength(), "\n \n " + Visits.get(selectedResident.getResID()), attrText);
+            docVisits.insertString(docVisits.getLength(), "\n \n " + factory.getSingleVisitDescription(selectedResident.getResID()), attrText);
 
         } catch (NullPointerException e) {
             System.out.println("NullPointerException");
@@ -499,18 +499,37 @@ public class GUI extends JFrame {
         }
     }
 
-    public void saveChangesResidentIncidentText(int index) {
-        String newText = taResident[index].getText();
-        int resID = DatabaseFactory.residents.get(index).getResID();
-        //todo datumsabhängigkeit
-        DatabaseService.updateIncidentsDatabase(newText, DatabaseFactory.incidents.get(index));
+    private void saveChangesResidentIncidentText(int index) {
+        String newText = taResident[index].getText();  //Get text that has been changed
+        int resID = factory.residents.get(index).getResID(); //get resid of selected resident
+        String dateString = (String) jcbTime.getSelectedItem(); //String format
+        Date date = null;
+        try {
+            date = new SimpleDateFormat("dd.MM.yyyy").parse(dateString);
+        } catch (ParseException e) {
+            e.printStackTrace();
+        }
+
+        factory.saveResidentIncidentsDatabase(newText, factory.getSingleIncident(resID,date), resID, date);
     }
 
+    //todo datumsabhängigkeit
+
     public void saveChangesShiftIncidentText() {
-        String newText = taAll.getText();
+        String newText = taAll.getText(); //get new text
+        String dateString = (String) jcbTime.getSelectedItem(); //String format
+        int shiftCategory = (jcbShift.getSelectedIndex()) + 1; //get selected shift category
+        Date date = null;
+        try {
+            date = new SimpleDateFormat("dd.MM.yyyy").parse(dateString);
+        } catch (ParseException e) {
+            e.printStackTrace();
+        }
+        int shiftID = factory.getSingleShiftSchedule(shiftCategory, date).getShiftID();
+        factory.saveShiftIncidentsDatabase(newText, shiftID);
         //todo datums und shiftabhängigkeit
-       // DatabaseService.updateIncidentsDatabase(newText, DatabaseFactory.incidents.get(index));
     }
+
 
     class ButtonListenerChangeCardsForResidentSpecificData implements ActionListener {
         @Override
